some changes in navigation
