+++ src/navigation/RootNavigation.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { NavigationContainer, StackActions } from '@react-navigation/native';
+import { StackNavigator } from './StackNavigator';
+export function RootNavigation(){
+    return (
+        <NavigationContainer>
+            <StackNavigator/>
+        </NavigationContainer>
+    )
+}
