+++ src/navigation/StackNavigator.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { createStackNavigator } from "@react-navigation/stack";
+import HomeScreen from "../screens/Home";
+
+const Stack=createStackNavigator()
+export function StackNavigator(){
+    return (
+        <Stack.Navigator>
+            <Stack.Screen name='Home' component={HomeScreen}/>
+        </Stack.Navigator>
+    )
+}
